+++ HISTORY.rst
@@ -7,7 +7,7 @@ History
 
 New indicators
 ~~~~~~~~~~~~~~
-* `land.snow_melt_we_max` computes the maximum snow melt over n days
+* `land.snow_melt_we_max` computes the maximum snow melt over n days, and `land.melt_and_precip_max` the maximum combined snow melt and precipitation.
 * `snd_max_doy` returns the day of the year where snow depth reaches its maximum value.
 * `land.snow_cover_duration` computes the number of days snow depth exceeds some minimal threshold.
 * `land.continuous_snow_cover_start` and `land.continuous_snow_cover_end` identify the day of the year when snow depth crosses a threshold for a given period of time.